requirements.txt
pyinstaller 
pynput 
keyboard 
pyautogui 
pystray 
pillow 
PyQt6 
